Changed how the secret is stored
Now instead of saving the secret in every file it is stored in one unique file. This drastically helps with 2 things. First of all the size of the generated folder is drastically smaller. Second of all, since the size on disk is way smaller, the program is much less I/O bound in both the encoding and the decoding aspects of the program, which makes it much faster than before